Updates namings in more parts
